split fitting engine logic
