+++ scripts/Engineering/texture/TextureUtils/multidomain_engine.py
@@ -1,0 +1,524 @@
+# Mantid Repository : https://github.com/mantidproject/mantid
+#
+# Copyright &copy; 2026 ISIS Rutherford Appleton Laboratory UKRI,
+#   NScD Oak Ridge National Laboratory, European Spallation Source,
+#   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
+# SPDX - License - Identifier: GPL - 3.0 +
+
+# MultiDomainFunction fitting engine for the texture peak-fitting workflow.
+#
+# The original iterative MultiDomainFunction fit: a summed-spectrum fit seeds the peak centre, then
+# each (workspace, peak) is fit serially on progressively finer rebunched copies for SNR, with the
+# shared broadening params tied across domains on the first fit and held fixed per-domain thereafter.
+# Fits in TOF.  Retained for comparison/validation and for its unweighted-cost behaviour.
+#
+# The shared seeding, cropping and small helpers live in fitting_utils; this module composes them.
+
+from typing import Sequence, Tuple, List
+
+import numpy as np
+
+from mantid.simpleapi import (
+    SaveNexus,
+    logger,
+    CreateEmptyTableWorkspace,
+    Fit,
+    ConvertUnits,
+    Rebunch,
+    CloneWorkspace,
+)
+from mantid.api import AnalysisDataService as ADS, MultiDomainFunction, FunctionFactory
+from mantid.dataobjects import Workspace2D
+
+from Engineering.EnggUtils import convert_TOFerror_to_derror
+from plugins.algorithms.IntegratePeaks1DProfile import get_eval_ws, calc_sigma_from_summation
+
+from .fitting_utils import (
+    _PEAK_INTENSITY_PARAM,
+    _PEAK_CENTRE_PARAM,
+    _param_names,
+    _d_to_tof,
+    _tof_to_d,
+    _fit_parameters_path,
+    _make_composite,
+    _estimate_intensity_background_and_centre,
+    _tie_bkg,
+    fit_initial_summed_spectra,
+    _get_run_and_prefix_from_ws_log,
+    _get_grouping_from_ws_log,
+    get_default_values,
+    replace_nans,
+)
+
+
+# ~~~~~~~~~~~~~~~~~~~~~ MultiDomainFunction per-spectrum fit setup ~~~~~~~~~~~~~~~~~~~~~
+
+
+def get_initial_fit_function_and_kwargs_from_specs(
+    ws: Workspace2D,
+    ws_tof: Workspace2D,
+    peak: float,
+    x_window: tuple[float, float],
+    x0_window: tuple[float, float],
+    parameters_to_tie: Sequence[str] | None,
+    peak_func_name: str,
+    bg_func_name: str,
+    tie_bkg: bool,
+) -> Tuple[str, dict, Sequence[float]]:
+    # modification of get_initial_fit_function_and_kwargs to just fit a peak within the x_window
+    # NB: the shared broadening params (parameters_to_tie) are tied across domains here so this
+    # initial fit establishes a single, unit-correct (TOF) value for them.  Subsequent fits
+    # (rerun_fit_with_new_ws) then hold that value fixed per-spectrum, decoupling the domains.
+
+    # get the number of spectra
+    si = ws.spectrumInfo()
+    ispecs = list(range(si.size()))
+
+    # set up the fit window and data structures
+    x_start, x_end = x_window
+    fit_kwargs = {}
+    approx_bkgs = []
+    intensity_estimates = []
+
+    # set up the overall fit wrapper function and base versions of the individual peak and background
+    function = MultiDomainFunction()
+    bg_func = FunctionFactory.createFunction(bg_func_name)
+    base_peak_func = FunctionFactory.Instance().createPeakFunction(peak_func_name)
+
+    # save the width parameter name for future constraints
+    width_par_name = base_peak_func.getWidthParameterName()
+
+    # for each of the spectra
+    for ispec in ispecs:
+        # convert d spacing to tof - better step size for A, B and S refinement
+        diff_consts = si.diffractometerConstants(ispec)
+        tof_peak = _d_to_tof(peak, diff_consts)
+        tof_start = _d_to_tof(x_start, diff_consts)
+        tof_end = _d_to_tof(x_end, diff_consts)
+
+        # get the window indices for the spectra
+        istart = ws_tof.yIndexOfX(tof_start, ispec)
+        iend = ws_tof.yIndexOfX(tof_end, ispec)
+
+        # get param estimates
+        intens, sigma, bg, centre = _estimate_intensity_background_and_centre(ws_tof, ispec, istart, iend, tof_peak)
+        intensity_estimates.append(intens)
+        approx_bkgs.append(bg)
+
+        # create an individual peak, using estimated values as initial guess
+        peak_func = FunctionFactory.Instance().createPeakFunction(peak_func_name)
+        peak_func.setParameter(_PEAK_CENTRE_PARAM, centre)
+        peak_func.setParameter(_PEAK_INTENSITY_PARAM, intens)
+        peak_func.setMatrixWorkspace(ws_tof, ispec, tof_start, tof_end)
+
+        # calculate constraint values
+        # convert x0 bounds to TOF
+        x0_lower = _d_to_tof(x0_window[0], diff_consts)
+        x0_upper = _d_to_tof(x0_window[1], diff_consts)
+        # constrain the values of S to be at least half bin width and no more than half the window size
+        scale_factor = 2 * np.sqrt(2 * np.log(2))
+        width_min = 0.5 * ((tof_end - tof_start) / (iend - istart)) * scale_factor
+        width_max = max(width_min + 1e-10, ((tof_end - tof_start) / 2) * scale_factor)
+
+        # add these constraints
+        peak_func.addConstraints(f"{x0_lower} < {_PEAK_CENTRE_PARAM} < {x0_upper}")
+        peak_func.addConstraints(f"0 < {_PEAK_INTENSITY_PARAM} < {intens * 5}")
+        if width_par_name:
+            peak_func.addConstraints(f"{width_min}<{width_par_name}<{width_max}")
+
+        if not tie_bkg:
+            bg_func.setParameter("A0", bg)
+
+        # package up the spectra fit functions (peak + background) into a composite function
+        comp_func = _make_composite(peak_func, bg_func)
+        function.add(comp_func)
+        function.setDomainIndex(ispec, ispec)
+        function.setMatrixWorkspace(ws_tof, ispec, tof_start, tof_end)
+
+        # set the fit kwargs for this spectra
+        key_suffix = f"_{ispec}" if ispec > 0 else ""
+        fit_kwargs["InputWorkspace" + key_suffix] = ws_tof.name()
+        fit_kwargs["StartX" + key_suffix] = tof_start
+        fit_kwargs["EndX" + key_suffix] = tof_end
+        fit_kwargs["WorkspaceIndex" + key_suffix] = int(ispec)
+
+    # add parameter ties
+    ties = []
+
+    # first tie background
+    if tie_bkg:
+        function, ties = _tie_bkg(function, approx_bkgs, ties)
+
+    # then any other nominated parameters
+    available_params = _param_names(base_peak_func)
+    if parameters_to_tie is not None:
+        invalid = [p for p in parameters_to_tie if p not in available_params]
+        if invalid:
+            raise ValueError(f"Invalid parameter(s) to tie: {invalid}. Available: {available_params}")
+        else:
+            for idom in range(1, function.nDomains()):
+                # tie global params to first
+                for par in parameters_to_tie:
+                    ties.append(f"f{idom}.f0.{par}=f0.f0.{par}")  # global peak pars
+
+    # add ties directly to function object to preserve workspace references on member functions
+    if ties:
+        function.addTies(",".join(ties))
+    return function, fit_kwargs, intensity_estimates
+
+
+def rerun_fit_with_new_ws(
+    mdf: MultiDomainFunction,
+    fit_kwargs: dict,
+    md_fit_kwargs: dict,
+    new_ws: Workspace2D,
+    x0_frac_move: float,
+    iters: int,
+    parameters_to_fix: Sequence[str] | None = None,
+    parameters_to_tie: Sequence[str] | None = None,
+    tie_background: bool = False,
+    is_final: bool = False,
+    last_fit_ic: bool = False,
+) -> Tuple[Fit, dict]:
+    # update the input workspace in the fitting kwargs
+    for k in md_fit_kwargs.keys():
+        if "InputWorkspace" in k:
+            md_fit_kwargs[k] = new_ws.name()
+
+    ties = []
+    new_func = MultiDomainFunction()
+    for idom in range(mdf.nFunctions()):
+        comp = mdf[idom]
+        peak = comp[0]
+        bg = comp[1]
+
+        peak_name = "IkedaCarpenterPV" if last_fit_ic and is_final else peak.name()
+
+        new_peak = FunctionFactory.Instance().createPeakFunction(peak_name)
+
+        intens = max(peak.getParameterValue(_PEAK_INTENSITY_PARAM), 1)
+        x0 = peak.getParameterValue(_PEAK_CENTRE_PARAM)
+        key_suffix = f"_{idom}" if idom > 0 else ""
+
+        # create fresh peak as ties are causing problems
+        if last_fit_ic and is_final:
+            # set X0 and I BEFORE setMatrixWorkspace so that the instrument
+            # parameter file formulas for SigmaSquared and Gamma are evaluated
+            # at the correct peak centre rather than at X0=0 (default)
+            new_peak.setParameter(_PEAK_CENTRE_PARAM, x0)
+            new_peak.setParameter(_PEAK_INTENSITY_PARAM, intens)
+            new_peak.setMatrixWorkspace(new_ws, idom, md_fit_kwargs["StartX" + key_suffix], md_fit_kwargs["EndX" + key_suffix])
+            # if we have changed to IC for the last fit we will just use default parameter ties
+            if peak.name() != "IkedaCarpenter":
+                parameters_to_tie = _get_default_param_ties("IkedaCarpenterPV", None)
+        else:
+            for param in _param_names(new_peak):
+                new_peak.setParameter(param, peak.getParameterValue(param))
+        # set workspace on peak function directly so IkedaCarpenterPV can calculate wavelengths
+        new_peak.setMatrixWorkspace(new_ws, idom, md_fit_kwargs["StartX" + key_suffix], md_fit_kwargs["EndX" + key_suffix])
+
+        # update constraints around new values
+
+        if not is_final:
+            # don't constrain the intensity on the final fit
+            new_peak.addConstraints(f"{max(intens / 2, 1e-6)}<{_PEAK_INTENSITY_PARAM}<{intens * 2}")
+        new_peak.addConstraints(f"{x0 * (1 - x0_frac_move)}<{_PEAK_CENTRE_PARAM}<{x0 * (1 + x0_frac_move)}")
+        # tie background across domains if requested (kept - a cheap linear-background coupling that
+        # helps weak spectra share a level); the shared broadening params are fixed per-domain below
+        # rather than tied, so each spectrum's peak shape fits independently
+        if idom > 0 and tie_background:
+            for ipar_bg in range(bg.nParams()):
+                par = bg.getParamName(ipar_bg)
+                ties.append(f"f{idom}.f1.{par}=f0.f1.{par}")
+        # fix user-nominated parameters
+        if parameters_to_fix:
+            for param in parameters_to_fix:
+                new_peak.fixParameter(param)
+        # Hold the shared broadening params fixed at the value carried through from the previous fit
+        # (the initial fit tied them across domains to establish one consistent value).  Fixing them
+        # here decouples the domains so the multi-domain solve is block-diagonal.  setMatrixWorkspace
+        # above can reapply the instrument parameter formulas, so re-assert the previous value before
+        # fixing - except for a freshly created IkedaCarpenterPV, whose params come from the instrument.
+        if parameters_to_tie:
+            fresh_ic = last_fit_ic and is_final
+            for par in parameters_to_tie:
+                if not fresh_ic:
+                    new_peak.setParameter(par, peak.getParameterValue(par))
+                new_peak.fixParameter(par)
+
+        comp_func = _make_composite(new_peak, bg)
+        new_func.add(comp_func)
+        new_func.setDomainIndex(idom, idom)
+
+    # add ties directly to function object to preserve workspace references on member functions
+    if ties:
+        new_func.addTies(",".join(ties))
+
+    return Fit(
+        Function=new_func,
+        Output=f"fit_{new_ws.name()}",
+        MaxIterations=iters,
+        **fit_kwargs,
+        **md_fit_kwargs,
+    ), md_fit_kwargs
+
+
+def _get_default_param_ties(peak_func_name: str, parameters_to_tie: Sequence[str] | None) -> Sequence[str]:
+    if not parameters_to_tie:
+        match peak_func_name:
+            case "BackToBackExponential":
+                parameters_to_tie = ("A", "B")
+            case "IkedaCarpenterPV":
+                parameters_to_tie = ("Alpha0", "Alpha1", "Beta0", "Kappa")
+    return parameters_to_tie
+
+
+def calc_intens_and_sigma_arrays(fit_result: Fit) -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    function = fit_result["Function"]
+    ndoms = function.nDomains()
+    intens = np.zeros(ndoms)
+    sigma = np.zeros(intens.shape)
+    intens_over_sig = np.zeros(intens.shape)
+    peak_limits = np.full(intens.shape, None)
+    for idom, comp_func in enumerate(function):
+        intens[idom] = comp_func.getParameterValue("f0.I")
+        ws_fit = get_eval_ws(fit_result["OutputWorkspace"], idom, ndoms)
+        sigma[idom], peak_limits[idom] = calc_sigma_from_summation(ws_fit.readX(0), ws_fit.readE(0) ** 2, ws_fit.readY(3))
+    ivalid = ~np.isclose(sigma, 0)
+    intens_over_sig[ivalid] = intens[ivalid] / sigma[ivalid]
+    return intens, sigma, intens_over_sig, peak_limits
+
+
+# ~~~~~~~~~~~~~~~~~~~~~~~~ MultiDomainFunction fitting engine ~~~~~~~~~~~~~~~~~~~~~~~~
+
+
+def _build_fit_ws_stack(
+    ws_tof: Workspace2D,
+    smooth_vals: Sequence[int],
+    tied_bkgs: Sequence[bool],
+    final_fit_raw: bool,
+    last_fit_ic: bool,
+    peak_func_name: str,
+) -> Tuple[List, List[bool]]:
+    """Build the ordered stack of workspaces to fit iteratively (progressively finer rebunched copies
+    for SNR, then optionally the raw ws, then an extra raw copy when the final fit switches to
+    IkedaCarpenterPV) together with the per-fit 'tie background' flags."""
+    fit_wss = []
+    bkg_is_tied = []
+    if len(smooth_vals) > 0:
+        for i, smooth_val in enumerate(smooth_vals):
+            fit_wss.append(Rebunch(InputWorkspace=ws_tof, OutputWorkspace=f"smooth_ws_{smooth_val}", NBunch=smooth_val))
+            bkg_is_tied.append(tied_bkgs[i])
+    # if no smoothing values are given, the initial fit should just be on ws_tof
+    # if final_fit_raw flagged, ws_tof should be added to the end of the fit_wss stack
+    if final_fit_raw or len(smooth_vals) == 0:
+        fit_wss.append(ws_tof)
+        bkg_is_tied.append(True)
+    # if the peak func isn't already Ikeda Carpenter, and last_fit_ic is true, add another ws_tof to be fit
+    if last_fit_ic and peak_func_name != "IkedaCarpenterPV":
+        fit_wss.append(CloneWorkspace(ws_tof, OutputWorkspace="ws_tof_IkedaCarpenter"))
+        bkg_is_tied.append(True)
+    return fit_wss, bkg_is_tied
+
+
+def _run_iterative_multidomain_fit(
+    ws: Workspace2D,
+    fit_wss: Sequence,
+    bkg_is_tied: Sequence[bool],
+    peak: float,
+    x_window: Tuple[float, float],
+    x0_window: Tuple[float, float],
+    parameters_to_tie: Sequence[str] | None,
+    subsequent_fit_param_fix: Sequence[str] | None,
+    peak_func_name: str,
+    bg_func_name: str,
+    fit_kwargs: dict,
+    max_fit_iters: int,
+    last_fit_ic: bool,
+) -> Tuple[Fit, Workspace2D, Sequence[float]]:
+    """Fit the first (coarsest) workspace in the stack, then re-fit on each subsequent workspace,
+    carrying the fitted function forward.  Returns the final Fit object, the last workspace fitted
+    (used to retrieve its parameter table), and the per-spectrum intensity estimates from the
+    initial fit setup."""
+    # perform initial fit set up and fit
+    fit_num = 0
+    fit_ws = fit_wss[fit_num]
+    initial_function, md_fit_kwargs, intensity_estimates = get_initial_fit_function_and_kwargs_from_specs(
+        ws, fit_ws, peak, x_window, x0_window, parameters_to_tie, peak_func_name, bg_func_name, bkg_is_tied[fit_num]
+    )
+    fit_object = Fit(
+        Function=initial_function,
+        Output=f"fit_{fit_ws}",
+        MaxIterations=max_fit_iters,
+        **fit_kwargs,
+        **md_fit_kwargs,
+    )
+
+    # perform subsequent fits
+    while len(fit_wss) - 1 > fit_num:
+        fit_num += 1
+        mdf = fit_object.Function.function
+        fit_ws = fit_wss[fit_num]
+        fit_object, md_fit_kwargs = rerun_fit_with_new_ws(
+            mdf,
+            fit_kwargs,
+            md_fit_kwargs,
+            fit_ws,
+            0.02,  # allow x0 to only vary by 2% from previous fit
+            max_fit_iters,
+            subsequent_fit_param_fix,
+            parameters_to_tie,
+            bkg_is_tied[fit_num],
+            fit_num == len(fit_wss) - 1,
+            last_fit_ic,
+        )
+    return fit_object, fit_ws, intensity_estimates
+
+
+def _populate_multidomain_output_table(
+    out_tab,
+    ws: Workspace2D,
+    fit_ws: Workspace2D,
+    intensity_estimates: Sequence[float],
+    fit_mask: np.ndarray,
+    no_fit_value_dict: dict | None,
+    nan_replacement: str | None,
+) -> None:
+    """Build the per-(ws, peak) output table (columns wsindex, I_est, then for each peak-function
+    parameter p: p, p_err, p/p_err) from the flat multidomain Fit parameter table produced for
+    fit_ws.  Spectra flagged False in fit_mask get default/unfit values.  X0 is converted from TOF
+    back to d-spacing per spectrum (the multidomain engine fits in TOF)."""
+    spec_fit = ADS.retrieve(f"fit_{fit_ws}_Parameters")
+    si = ws.spectrumInfo()
+    out_tab.addColumn("int", "wsindex")
+    out_tab.addColumn("double", "I_est")
+    all_params = spec_fit.column("Name")[:-1]  # last row is cost function
+    param_vals = spec_fit.column("Value")[:-1]
+    param_errs = spec_fit.column("Error")[:-1]
+    u_params = []
+    for col in all_params:  # last col is cost of whole fit
+        spec_num, func_num, param = col.split(".")
+        # assume first function is the peak
+        if func_num == "f0" and param not in u_params:
+            u_params.append(param)
+            out_tab.addColumn("double", param)
+            out_tab.addColumn("double", f"{param}_err")
+            out_tab.addColumn("double", f"{param}/{param}_err")
+
+    # get user defined default vals for unsuccessful fit parameters
+    default_vals = get_default_values(u_params, no_fit_value_dict)
+
+    # populate the rows of the table
+    table_vals = np.zeros((si.size(), 3 * len(u_params) + 1))  # intens_est + p_1_val, p_1_err, pm_1_val/err +...
+    for ispec in range(si.size()):
+        # logic for spectra which HAVE been fit successfully
+        if fit_mask[ispec]:
+            # add estimate of I
+            row = [intensity_estimates[ispec]]
+            for p in u_params:
+                param_name = f"f{ispec}.f0.{p}"
+                pind = all_params.index(param_name)
+                if p != _PEAK_CENTRE_PARAM:
+                    row += [param_vals[pind], param_errs[pind], np.divide(param_vals[pind], param_errs[pind])]
+                else:
+                    # for x0, convert back to d spacing
+                    diff_consts = si.diffractometerConstants(ispec)
+                    d_peak = _tof_to_d(param_vals[pind], diff_consts)
+                    d_err = convert_TOFerror_to_derror(diff_consts, param_errs[pind], d_peak)
+                    row += [d_peak, d_err, np.divide(d_peak, d_err)]
+        # logic for spectra which HAVE NOT been fit successfully
+        else:
+            row = [default_vals.get("I_est", np.nan)]
+            for p in u_params:
+                row += [default_vals[p], np.inf, 0.0]
+        table_vals[ispec] = row
+    if nan_replacement:
+        table_vals = replace_nans(table_vals, nan_replacement)
+    for i, row in enumerate(table_vals):
+        out_tab.addRow([i] + list(row))
+
+
+def _fit_all_peaks_multidomain(
+    wss: Sequence[str],
+    peaks: Sequence[float],
+    peak_window: float,
+    save_dir: str,
+    override_dir: bool,
+    i_over_sigma_thresh: float,
+    nan_replacement: str | None,
+    no_fit_value_dict: dict | None,
+    smooth_vals: Sequence[int],
+    tied_bkgs: Sequence[bool],
+    final_fit_raw: bool,
+    parameters_to_tie: Sequence[str] | None,
+    subsequent_fit_param_fix: Sequence[str] | None,
+    peak_func_name: str,
+    bg_func_name: str,
+    last_fit_ic: bool,
+    max_fit_iters: int,
+    fit_kwargs: dict,
+) -> None:
+    """MultiDomainFunction-based (serial, per (ws, peak)) implementation of fit_all_peaks.
+
+    The original iterative fit: a summed-spectrum fit seeds the peak centre, then each (ws, peak) is
+    fit on progressively finer rebunched copies for SNR (tied-then-fixed broadening) before the final
+    raw fit.  Fits in TOF.  Retained for comparison/validation and for the unweighted-cost behaviour."""
+    parameters_to_tie = _get_default_param_ties(peak_func_name, parameters_to_tie)
+
+    # we are initially going to fit a summed spectra to get a good starting point for the peak centre
+    # we will then fix the amount this can change in the individual fits
+    x0_lims, all_cropped_rebinned_wss = fit_initial_summed_spectra(wss, peaks, peak_window, fit_kwargs.copy(), peak_func_name)
+    for iws, wsname in enumerate(wss):
+        # notice user how far through the fitting they are (useful if any fits fail)
+        logger.notice(f"Fitting Workspace: {wsname} ({iws + 1}/{len(wss)})")
+
+        # obtain the ws and metadata about ws
+        ws = ADS.retrieve(wsname)
+        run, prefix = _get_run_and_prefix_from_ws_log(ws, wsname)
+        grouping = _get_grouping_from_ws_log(ws)
+
+        # loop over the peaks
+        for ipeak, peak in enumerate(peaks):
+            # perform fitting in TOF as the parameter magnitudes are better for fitting
+            ws_tof = ConvertUnits(InputWorkspace=all_cropped_rebinned_wss[ipeak][iws], OutputWorkspace="ws_tof", Target="TOF")
+
+            # approach will be to use iterative fits and these iterations can have optionally 'rebunched' data to improve SNR
+            fit_wss, bkg_is_tied = _build_fit_ws_stack(ws_tof, smooth_vals, tied_bkgs, final_fit_raw, last_fit_ic, peak_func_name)
+
+            # low level information
+            logger.information(f"Workspace: {wsname}, Peak: {peak}")
+
+            # set up window bounds and the output table / file path
+            xmin, xmax = peak - peak_window, peak + peak_window
+            out_ws = f"{prefix}{run}_{peak}_{grouping}_Fit_Parameters"
+            out_file = out_ws + ".nxs"
+            out_path = _fit_parameters_path(save_dir, override_dir, grouping, peak, out_file)
+            out_tab = CreateEmptyTableWorkspace(OutputWorkspace=out_ws)
+
+            # run the initial fit and any subsequent (finer / raw / IC) fits across the ws stack
+            fit_object, fit_ws, intensity_estimates = _run_iterative_multidomain_fit(
+                ws,
+                fit_wss,
+                bkg_is_tied,
+                peak,
+                (xmin, xmax),
+                x0_lims[ipeak],
+                parameters_to_tie,
+                subsequent_fit_param_fix,
+                peak_func_name,
+                bg_func_name,
+                fit_kwargs,
+                max_fit_iters,
+                last_fit_ic,
+            )
+
+            # establish which detectors have sufficient I over sigma
+            mdf = fit_object.Function.function
+            fit_result = {"Function": mdf, "OutputWorkspace": fit_object.OutputWorkspace.name()}
+            # update peak mask based on I/sig from fit
+            *_, i_over_sigma, _ = calc_intens_and_sigma_arrays(fit_result)
+            fit_mask = i_over_sigma > i_over_sigma_thresh
+
+            # populate and save the final table
+            _populate_multidomain_output_table(out_tab, ws, fit_ws, intensity_estimates, fit_mask, no_fit_value_dict, nan_replacement)
+            SaveNexus(InputWorkspace=out_ws, Filename=out_path)
